requirements.txt
opencv-python
mediapipe==0.10.9
numpy<2.0.0
pywin32
protobuf==3.20.3
msvc-runtime
Pillow